Misc fixes based on pull request feedback
